+++ data/src/stream.rs
@@ -45,17 +45,22 @@ pub enum Update {
         is_initial: bool,
         error: Option<String>,
         sent_time: DateTime<Utc>,
+        autoconnect: bool,
     },
     ConnectionFailed {
         server: Server,
         error: String,
         sent_time: DateTime<Utc>,
+        autoconnect: bool,
     },
     MessagesReceived(Server, Vec<message::Encoded>),
     Remove(Server),
     UpdateConfiguration {
         server: Server,
         updated_config: Arc<config::Server>,
+    },
+    AutoconnectDisabled {
+        server: Server,
     },
 }
 
@@ -84,8 +89,15 @@ enum Input {
 }
 
 pub enum Control {
-    Disconnect(Option<String>),
-    Connect,
+    Disconnect {
+        error: Option<String>,
+        disable_autoconnect: bool,
+    },
+    AuthenticationFailed {
+        error: Option<String>,
+    },
+    Connect(bool),
+    DisableAutoconnect,
     End(Option<String>),
     UpdateConfiguration(Arc<config::Server>, Option<config::Proxy>),
 }
@@ -124,6 +136,11 @@ async fn _run(
     });
 
     let mut is_initial = true;
+
+    // Needs to be tracked across states as connection can fail during
+    // authentication
+    let mut connection_attempt = 0;
+
     let mut state = State::Disconnected {
         autoconnect: config.autoconnect,
         retry: time::interval(config.reconnect_delay),
@@ -135,6 +152,7 @@ async fn _run(
         is_initial,
         error: None,
         sent_time: Utc::now(),
+        autoconnect: config.autoconnect,
     });
 
     loop {
@@ -147,7 +165,7 @@ async fn _run(
                             retry
                                 .tick()
                                 .into_stream()
-                                .map(|_| Control::Connect)
+                                .map(|_| Control::Connect(true))
                                 .boxed(),
                         )
                         .next()
@@ -170,10 +188,31 @@ async fn _run(
                             retry: time::interval(config.reconnect_delay),
                         };
                     }
-                    Some(Control::Disconnect(_)) => {
-                        *autoconnect = false;
+                    Some(Control::Disconnect {
+                        disable_autoconnect,
+                        ..
+                    }) => {
+                        if disable_autoconnect {
+                            *autoconnect = false;
+                        }
                     }
-                    Some(Control::Connect) => {
+                    Some(Control::DisableAutoconnect) => {
+                        *autoconnect = false;
+
+                        let _ = sender.unbounded_send(
+                            Update::AutoconnectDisabled {
+                                server: server.clone(),
+                            },
+                        );
+                    }
+                    Some(Control::Connect(automated)) => {
+                        if automated {
+                            connection_attempt += 1;
+                        } else {
+                            *autoconnect = config.autoconnect;
+                            connection_attempt = 1;
+                        }
+
                         let _ = sender.unbounded_send(Update::Connecting {
                             server: server.clone(),
                             sent_time: Utc::now(),
@@ -226,22 +265,29 @@ async fn _run(
                                     "[{server}] connection failed: {error}"
                                 );
 
+                                retry.reset();
+
+                                if connection_attempt
+                                    >= config.max_connection_attempts
+                                {
+                                    *autoconnect = false;
+                                }
+
                                 let _ = sender.unbounded_send(
                                     Update::ConnectionFailed {
                                         server: server.clone(),
                                         error,
                                         sent_time: Utc::now(),
+                                        autoconnect: *autoconnect,
                                     },
                                 );
-
-                                retry.reset();
                             }
                         }
                     }
                     Some(Control::End(_)) => {
                         state = State::End;
                     }
-                    None => (),
+                    Some(Control::AuthenticationFailed { .. }) | None => (),
                 }
             }
             State::Connected {
@@ -280,7 +326,12 @@ async fn _run(
                                 *requested_at + QUIT_REQUEST_TIMEOUT,
                             )
                             .into_stream()
-                            .map(|()| Input::Control(Control::Disconnect(None)))
+                            .map(|()| {
+                                Input::Control(Control::Disconnect {
+                                    error: None,
+                                    disable_autoconnect: true,
+                                })
+                            })
                             .boxed(),
                         );
                     }
@@ -304,6 +355,9 @@ async fn _run(
                             *ping_timeout = None;
                         }
                         proto::Command::ERROR(error) => {
+                            let autoconnect = !quit_requested.is_some();
+                            connection_attempt = 0;
+
                             if quit_requested.is_some() {
                                 let _ = sender.unbounded_send(
                                     Update::Disconnected {
@@ -311,6 +365,7 @@ async fn _run(
                                         is_initial,
                                         error: None,
                                         sent_time: Utc::now(),
+                                        autoconnect,
                                     },
                                 );
 
@@ -318,7 +373,7 @@ async fn _run(
                                 // a valid acknowledgement
                                 // https://modern.ircdocs.horse/#quit-message
                                 state = State::Disconnected {
-                                    autoconnect: false,
+                                    autoconnect,
                                     retry: time::interval_at(
                                         Instant::now() + config.reconnect_delay,
                                         config.reconnect_delay,
@@ -326,16 +381,18 @@ async fn _run(
                                 };
                             } else {
                                 log::info!("[{server}] disconnected: {error}");
+
                                 let _ = sender.unbounded_send(
                                     Update::Disconnected {
                                         server: server.clone(),
                                         is_initial,
                                         error: Some(error),
                                         sent_time: Utc::now(),
+                                        autoconnect,
                                     },
                                 );
                                 state = State::Disconnected {
-                                    autoconnect: true,
+                                    autoconnect,
                                     retry: time::interval_at(
                                         Instant::now() + config.reconnect_delay,
                                         config.reconnect_delay,
@@ -352,14 +409,19 @@ async fn _run(
                     }
                     Input::IrcMessage(Err(e)) => {
                         log::info!("[{server}] disconnected: {e}");
+
+                        let autoconnect = quit_requested.is_none();
+                        connection_attempt = 0;
+
                         let _ = sender.unbounded_send(Update::Disconnected {
                             server: server.clone(),
                             is_initial,
                             error: Some(e.to_string()),
                             sent_time: Utc::now(),
+                            autoconnect,
                         });
                         state = State::Disconnected {
-                            autoconnect: quit_requested.is_none(),
+                            autoconnect,
                             retry: time::interval_at(
                                 Instant::now() + config.reconnect_delay,
                                 config.reconnect_delay,
@@ -401,14 +463,19 @@ async fn _run(
                     }
                     Input::PingTimeout => {
                         log::info!("[{server}] ping timeout");
+
+                        let autoconnect = quit_requested.is_none();
+                        connection_attempt = 0;
+
                         let _ = sender.unbounded_send(Update::Disconnected {
                             server: server.clone(),
                             is_initial,
                             error: Some("ping timeout".into()),
                             sent_time: Utc::now(),
+                            autoconnect,
                         });
                         state = State::Disconnected {
-                            autoconnect: quit_requested.is_none(),
+                            autoconnect,
                             retry: time::interval_at(
                                 Instant::now() + config.reconnect_delay,
                                 config.reconnect_delay,
@@ -421,43 +488,20 @@ async fn _run(
                             updated_default_proxy,
                         ) => {
                             // If connection detail(s) change, then disconnect
-                            if config.server != updated_config.server
-                                || config.port != updated_config.port
-                                || config.use_tls != updated_config.use_tls
-                                || config.use_websocket
-                                    != updated_config.use_websocket
-                                || config.websocket_path
-                                    != updated_config.websocket_path
-                                || config.dangerously_accept_invalid_certs
-                                    != updated_config
-                                        .dangerously_accept_invalid_certs
-                                || config.root_cert_path
-                                    != updated_config.root_cert_path
-                                || config
-                                    .proxy
-                                    .as_ref()
-                                    .or(default_proxy.as_ref())
-                                    != updated_config
-                                        .proxy
-                                        .as_ref()
-                                        .or(updated_default_proxy.as_ref())
-                                || config.username != updated_config.username
-                                || config.password != updated_config.password
-                                || config.password_file
-                                    != updated_config.password_file
-                                || config.password_file_first_line_only
-                                    != updated_config
-                                        .password_file_first_line_only
-                                || config.password_command
-                                    != updated_config.password_command
-                                || config.sasl != updated_config.sasl
-                            {
+                            if updated_config.has_same_connection_settings(
+                                updated_default_proxy.as_ref(),
+                                &config,
+                                default_proxy.as_ref(),
+                            ) {
+                                connection_attempt = 0;
+
                                 let _ = sender.unbounded_send(
                                     Update::Disconnected {
                                         server: server.clone(),
                                         is_initial,
                                         error: None,
                                         sent_time: Utc::now(),
+                                        autoconnect: updated_config.autoconnect,
                                     },
                                 );
                                 state = State::Disconnected {
@@ -479,17 +523,53 @@ async fn _run(
                             config = updated_config;
                             default_proxy = updated_default_proxy;
                         }
-                        Control::Connect => (),
-                        Control::Disconnect(error) => {
+                        Control::Connect(_) | Control::DisableAutoconnect => (),
+                        Control::AuthenticationFailed { error } => {
+                            let autoconnect = if connection_attempt
+                                >= config.max_connection_attempts
+                            {
+                                false
+                            } else {
+                                config.autoconnect
+                            };
+
                             let _ =
                                 sender.unbounded_send(Update::Disconnected {
                                     server: server.clone(),
                                     is_initial,
                                     error,
                                     sent_time: Utc::now(),
+                                    autoconnect,
                                 });
                             state = State::Disconnected {
-                                autoconnect: false,
+                                autoconnect,
+                                retry: time::interval_at(
+                                    Instant::now() + config.reconnect_delay,
+                                    config.reconnect_delay,
+                                ),
+                            };
+                        }
+                        Control::Disconnect {
+                            error,
+                            disable_autoconnect,
+                        } => {
+                            let autoconnect = if disable_autoconnect {
+                                false
+                            } else {
+                                config.autoconnect
+                            };
+                            connection_attempt = 0;
+
+                            let _ =
+                                sender.unbounded_send(Update::Disconnected {
+                                    server: server.clone(),
+                                    is_initial,
+                                    error,
+                                    sent_time: Utc::now(),
+                                    autoconnect,
+                                });
+                            state = State::Disconnected {
+                                autoconnect,
                                 retry: time::interval_at(
                                     Instant::now() + config.reconnect_delay,
                                     config.reconnect_delay,
@@ -622,15 +702,26 @@ impl Map {
         }
     }
 
-    pub fn disconnect(&mut self, server: &Server, error: Option<String>) {
+    pub fn authentication_failed(
+        &mut self,
+        server: &Server,
+        error: Option<String>,
+    ) {
         if let Some(controller) = self.0.get_mut(server) {
-            let _ = controller.try_send(Control::Disconnect(error));
+            let _ =
+                controller.try_send(Control::AuthenticationFailed { error });
         }
     }
 
     pub fn connect(&mut self, server: &Server) {
         if let Some(controller) = self.0.get_mut(server) {
-            let _ = controller.try_send(Control::Connect);
+            let _ = controller.try_send(Control::Connect(false));
+        }
+    }
+
+    pub fn disable_autoconnect(&mut self, server: &Server) {
+        if let Some(controller) = self.0.get_mut(server) {
+            let _ = controller.try_send(Control::DisableAutoconnect);
         }
     }